Add setup script to prepare the emulator
There are several steps with a fresh system that need to be performed to
prepare for running the emulator. This adds a setup.sh script for Mac
and Linux systems that takes care of putting all of the Redfish and
Swordfish files in the right place and getting the necessary
requirements installed to the point of being able to run the emulator.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>